depr: deprecate `utils.downloads` in favor of `Dataset.download()` (#1030)
