+++ src/pymovements/dataset/dataset_download.py
@@ -24,9 +24,9 @@ import shutil
 from urllib.error import URLError
 
 from pymovements.dataset._utils._archives import extract_archive
+from pymovements.dataset._utils._downloads import download_file
 from pymovements.dataset.dataset_definition import DatasetDefinition
 from pymovements.dataset.dataset_paths import DatasetPaths
-from pymovements.utils.downloads import download_file
 
 
 def download_dataset(
